Update search config
Currently there is are 2 indexes containing duplicate data.
Logged-in eager index is to server employee users
By combining logged-in with price-admin, all price-admins will be
served by this index, as it combines more groups than just
'logged-in'.
Once supported by mu-search it would be more optimal to put
price-admin as a separate group to be ignored.